zad1/requirements.txt
cycler==0.10.0
docutils==0.16
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
pathlib==1.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
statistics==1.0.3.5
